[Enhancement] introduce fe extensions
StarRocks currently features a plugin mechanism that enables users to install and manage plugins via SQL, such as audit log plugins or data integration plugins.
However, the existing plugin mechanism primarily focuses on plugin installation/uninstallation and single-function implementation. It lacks the capability for deep extension into internal system behaviors and cannot uniformly manage plugin lifecycles, dependencies, and resources. This makes it difficult to meet more complex business scenarios and expansion requirements.
