+++ fe/fe-core/src/main/java/com/starrocks/StarRocksFEServer.java
@@ -43,6 +43,7 @@ import com.starrocks.common.ThreadPoolManager;
 import com.starrocks.common.Version;
 import com.starrocks.common.util.NetUtils;
 import com.starrocks.common.util.Util;
+import com.starrocks.extension.ExtensionManager;
 import com.starrocks.failpoint.FailPoint;
 import com.starrocks.ha.FrontendNodeType;
 import com.starrocks.ha.StateChangeExecutor;
@@ -124,6 +125,8 @@ public class StarRocksFEServer {
 
             // set dns cache ttl
             java.security.Security.setProperty("networkaddress.cache.ttl", String.valueOf(Config.dns_cache_ttl_seconds));
+
+            ExtensionManager.getInstance().loadExtensionsFromDir(Config.ext_dir);
 
             RestoreClusterSnapshotMgr.init(starRocksDir + "/conf/cluster_snapshot.yaml", cmdLineOpts.isStartFromSnapshot());
