+++ fe/fe-core/src/main/java/com/starrocks/persist/gson/internal/MapTypeDeserializer.java
@@ -1,0 +1,38 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.starrocks.persist.gson.internal;
+
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParseException;
+import com.starrocks.persist.gson.GsonUtils;
+import com.starrocks.type.MapType;
+
+// Todo: remove it after remove selectedFields
+public class MapTypeDeserializer implements JsonDeserializer<MapType> {
+    @Override
+    public MapType deserialize(JsonElement jsonElement, java.lang.reflect.Type type,
+                               JsonDeserializationContext jsonDeserializationContext)
+            throws JsonParseException {
+        JsonObject dumpJsonObject = jsonElement.getAsJsonObject();
+        JsonObject key = dumpJsonObject.getAsJsonObject("keyType");
+        com.starrocks.type.Type keyType = GsonUtils.GSON.fromJson(key, com.starrocks.type.Type.class);
+        JsonObject value = dumpJsonObject.getAsJsonObject("valueType");
+        com.starrocks.type.Type valueType = GsonUtils.GSON.fromJson(value, com.starrocks.type.Type.class);
+        return new MapType(keyType, valueType);
+    }
+}
