+++ fe/fe-core/src/main/java/com/starrocks/persist/gson/internal/PrimitiveTypeDeserializer.java
@@ -1,0 +1,35 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.starrocks.persist.gson.internal;
+
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonParseException;
+import com.starrocks.type.PrimitiveType;
+
+import java.lang.reflect.Type;
+
+public class PrimitiveTypeDeserializer implements JsonDeserializer<PrimitiveType> {
+    @Override
+    public PrimitiveType deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
+            throws JsonParseException {
+        try {
+            return PrimitiveType.valueOf(json.getAsJsonPrimitive().getAsString());
+        } catch (Throwable t) {
+            return PrimitiveType.INVALID_TYPE;
+        }
+    }
+}
